Updating show for similar entities
