+++ pages/2_Similar_Entities.py
@@ -2,12 +2,12 @@ import streamlit as st
 
 from sibylapp.view.utils import setup, filtering
 from sibylapp.view import similar_entities
-from sibylapp.compute import context
 
 
 setup.setup_page()
 
 # Sidebar ------------------------------------
 filtering.view_entity_select()
+filtering.view_filtering()
 
 similar_entities.view(st.session_state["eid"])
