+++ sibylapp/view/feature_importance.py
@@ -1,4 +1,6 @@
 import streamlit as st
+
+import sibylapp.view.utils.filtering
 from sibylapp.compute import importance
 from sibylapp.view.utils import helpers
 from sibylapp.compute.context import get_term
@@ -36,6 +38,6 @@ def show_table(df):
 def view():
     to_show = format_importance_to_view(importance.compute_importance())
     to_show = to_show.sort_values(by="Importance Value", axis="index", ascending=False)
-    to_show = helpers.process_options(to_show)
+    to_show = sibylapp.view.utils.filtering.process_options(to_show)
     show_table(to_show)
     return to_show["Feature"]
