+++ sibylapp/view/utils/helpers.py
@@ -3,6 +3,8 @@ from sibylapp.config import BAR_LENGTH, FLIP_COLORS
 from sibylapp.compute.context import get_term
 import math
 
+from sibylapp.view.utils.filtering import process_options
+
 if FLIP_COLORS:
     POS_EM = "🟥"
     NEG_EM = "🟦"
@@ -13,30 +15,6 @@ NEUT_EM = "🟪"
 BLANK_EM = "⬜"
 UP_ARROW = "⬆"
 DOWN_ARROW = "⬇"
-
-
-def process_options(to_show):
-    return process_search(process_filter(process_show_more(to_show)))
-
-
-def process_show_more(to_show):
-    if not st.session_state["show_more"]:
-        return to_show[0:10]
-    return to_show
-
-
-def process_search(to_show):
-    if st.session_state["search_term"] is not None:
-        to_show = to_show[
-            to_show["Feature"].str.contains(st.session_state["search_term"], case=False)
-        ]
-    return to_show
-
-
-def process_filter(to_show):
-    if len(st.session_state["filters"]) > 0:
-        return to_show[to_show["Category"].isin(st.session_state["filters"])]
-    return to_show
 
 
 def generate_bars(values, neutral=False):
